Add props files for inputs
