+++ packages/experiments/src/components/Form/inputs/FormCheckBox.Props.ts
@@ -1,0 +1,8 @@
+import { IFormBaseInputProps } from '../FormBaseInput';
+import { ICheckboxProps } from 'office-ui-fabric-react/lib/Checkbox';
+export { ICheckboxProps };
+
+export interface IFormCheckBoxProps extends IFormBaseInputProps<boolean> {
+  /** Extra props for the Fabric checkbox */
+  checkboxProps?: ICheckboxProps;
+}
