+++ packages/experiments/src/components/Form/inputs/FormDatePicker.Props.ts
@@ -1,0 +1,8 @@
+import { IDatePickerProps } from 'office-ui-fabric-react/lib/DatePicker';
+import { IFormBaseInputProps } from '../FormBaseInput';
+export { IDatePickerProps };
+
+export interface IFormDatePickerProps extends IFormBaseInputProps<Date> {
+  /** Props for the Fabric date picker */
+  datePickerProps?: IDatePickerProps;
+}
