+++ packages/experiments/src/components/Form/inputs/FormDropdown.Props.ts
@@ -1,0 +1,8 @@
+import { IDropdownProps } from 'office-ui-fabric-react/lib/Dropdown';
+import { IFormBaseInputProps } from '../FormBaseInput';
+export { IDropdownProps };
+
+export interface IFormDropdownProps extends IFormBaseInputProps<number | string> {
+  /** Props for the Dropdown component */
+  dropdownProps?: IDropdownProps;
+}
