+++ packages/experiments/src/components/Form/inputs/FormTextInput.Props.ts
@@ -1,0 +1,10 @@
+
+import { IFormBaseInputProps } from '../FormBaseInput';
+import { ITextFieldProps } from 'office-ui-fabric-react/lib/TextField';
+
+export { ITextFieldProps };
+
+export interface IFormTextInputProps extends IFormBaseInputProps<string> {
+  /** Props for the TextField component */
+  textFieldProps?: ITextFieldProps;
+}
